Flask/requirements.txt
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cors==1.0.1
dnspython==2.2.0
filelock==3.3.2
Flask==2.0.2
Flask-Cors==3.0.10
flask-unittest==0.1.2
future==0.18.2
gevent==21.8.0
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pymongo==3.12.1
pyparsing==3.0.8
PySocks==1.7.1
pyspf==2.0.14
pytest==7.1.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-file==1.5.1
six==1.16.0
tldextract==3.1.2
tomli==2.0.1
urllib3==1.26.7
Werkzeug==2.0.2
zope.event==4.5.0
zope.interface==5.4.0